breaking rule def off, importing statsheets to calc

+++ src/damage_calculator/mod.rs
@@ -9,11 +9,11 @@ use crate::{
 
 #[derive(Debug)]
 pub enum ReactionEffect {
-    Additive(f64),
-    Multiplicative(f64),
+    Additive(f32),
+    Multiplicative(f32),
 }
 impl ReactionEffect {
-    pub fn map(&self, f: impl Fn(f64) -> f64) -> Self {
+    pub fn map(&self, f: impl Fn(f32) -> f32) -> Self {
         match *self {
             Self::Additive(x) => Self::Additive(f(x)),
             Self::Multiplicative(x) => Self::Multiplicative(f(x)),
@@ -34,7 +34,7 @@ pub fn evaluate_damage_instance(
     target_stats: &StatSheet,
     target_aura: Option<&GaugedAura>,
     crit_mode: CritMode,
-) -> f64 {
+) -> f32 {
     let conds = talent.conditions_met();
     let mut result = base_dmg(stats, talent)
         * stats.sum_mults(conds.iter().map(|&x| BaseDMGMult(x)))
@@ -52,7 +52,7 @@ pub fn evaluate_damage_instance(
 }
 
 /// Calculates the base damage given the stats and talent
-pub fn base_dmg(stats: &StatSheet, talent: &Talent) -> f64 {
+pub fn base_dmg(stats: &StatSheet, talent: &Talent) -> f32 {
     let result = talent
         .get_scalings()
         .iter()
@@ -63,14 +63,14 @@ pub fn base_dmg(stats: &StatSheet, talent: &Talent) -> f64 {
 }
 
 /// Calculates the defense multiplier of the enemy
-pub fn def_mult(char_lvl: f64, enemy_lvl: f64) -> f64 {
+pub fn def_mult(char_lvl: f32, enemy_lvl: f32) -> f32 {
     let result = (char_lvl + 100.0) / (char_lvl + enemy_lvl + 200.0);
     println!("def mult: {result}");
     result
 }
 
 /// Calculates the resistance multiplier of the target
-pub fn res_mult(target_res: f64) -> f64 {
+pub fn res_mult(target_res: f32) -> f32 {
     let result = if target_res < 0.0 {
         1.0 - target_res / 2.0
     } else if target_res < 0.75 {
@@ -83,7 +83,7 @@ pub fn res_mult(target_res: f64) -> f64 {
 }
 
 /// Calculates the crit multiplier
-pub fn crit_mult(cr: f64, cd: f64, mode: CritMode) -> f64 {
+pub fn crit_mult(cr: f32, cd: f32, mode: CritMode) -> f32 {
     let result = match mode {
         CritMode::NonCrit => 1.0,
         CritMode::AvgCrit => 1.0 + cr * cd,
@@ -121,7 +121,7 @@ pub fn rxn_effect(
     result
 }
 
-pub fn rxn_em_mult(rxn: ElementalReaction, em: f64) -> f64 {
+pub fn rxn_em_mult(rxn: ElementalReaction, em: f32) -> f32 {
     match rxn {
         ForwardMelt | ForwardVaporize | ReverseMelt | ReverseVaporize => amp_em_mult(em),
         Aggravate | Spread => add_em_mult(em),
@@ -130,30 +130,30 @@ pub fn rxn_em_mult(rxn: ElementalReaction, em: f64) -> f64 {
     }
 }
 
-pub fn amp_em_mult(em: f64) -> f64 {
+pub fn amp_em_mult(em: f32) -> f32 {
     let result = 2.778 * em / (em + 1400.0);
     println!("amplifying em mult: {result}");
     result
 }
 
-pub fn trans_em_mult(em: f64) -> f64 {
+pub fn trans_em_mult(em: f32) -> f32 {
     let result = 16.0 * em / (em + 2000.0);
     println!("transformative em mult: {result}");
     result
 }
 
-pub fn add_em_mult(em: f64) -> f64 {
+pub fn add_em_mult(em: f32) -> f32 {
     let result = 5.0 * em / (em + 1200.0);
     println!("additive em mult: {result}");
     result
 }
 
-pub fn cry_em_dmg_absorb(em: f64) -> f64 {
+pub fn cry_em_dmg_absorb(em: f32) -> f32 {
     let result = 4.44 * em / (em + 1400.0);
     println!("crystallize em mult: {result}");
     result
 }
 
-pub fn level_multiplier(_level: f64) -> f64 {
+pub fn level_multiplier(_level: f32) -> f32 {
     todo!("level multipliers")
 }
